fix: cache keys inconsistent (#199)
* fix: cache keys inconsistent

* fix: cache keys inconsistents

* chore: lint fix

* chore: lint fix

+++ src/Listeners/Caches/GuildEmojis/GuildEmojisUpdate.ts
@@ -8,22 +8,23 @@ import { GenKey } from "../../../Utilities/GenKey.js";
 export class GuildEmojisUpdate extends Listener {
     public constructor(context: ListenerContext) {
         super(context, {
-            event: GatewayDispatchEvents.GuildEmojisUpdate,
-            enabled: stateEmojis
+            event: GatewayDispatchEvents.GuildEmojisUpdate
         });
     }
 
     public async run(payload: { data: GatewayGuildEmojisUpdateDispatch; shardId: number }): Promise<void> {
         const old_emojis = await redisSScanStreamPromise(this.store.redis, GenKey(`${RedisKey.EMOJI_KEY}${RedisKey.KEYS_SUFFIX}`, payload.data.d.guild_id), "*", 1000);
-        for (const emoji of old_emojis) {
-            await this.store.redis.unlink(emoji);
-            await this.store.redis.srem(GenKey(`${RedisKey.EMOJI_KEY}${RedisKey.KEYS_SUFFIX}`, payload.data.d.guild_id), emoji);
-        }
+        if (stateEmojis) {
+            for (const emoji of old_emojis) {
+                await this.store.redis.unlink(emoji);
+                await this.store.redis.srem(GenKey(`${RedisKey.EMOJI_KEY}${RedisKey.KEYS_SUFFIX}`, payload.data.d.guild_id), emoji);
+            }
 
-        for (const emoji of payload.data.d.emojis) {
-            if (emoji.id) {
-                await this.store.redis.set(GenKey(RedisKey.EMOJI_KEY, emoji.id, payload.data.d.guild_id), JSON.stringify(emoji));
-                await this.store.redis.sadd(GenKey(`${RedisKey.EMOJI_KEY}${RedisKey.KEYS_SUFFIX}`, payload.data.d.guild_id), GenKey(RedisKey.EMOJI_KEY, emoji.id, payload.data.d.guild_id));
+            for (const emoji of payload.data.d.emojis) {
+                if (emoji.id) {
+                    await this.store.redis.set(GenKey(RedisKey.EMOJI_KEY, emoji.id, payload.data.d.guild_id), JSON.stringify(emoji));
+                    await this.store.redis.sadd(GenKey(`${RedisKey.EMOJI_KEY}${RedisKey.KEYS_SUFFIX}`, payload.data.d.guild_id), GenKey(RedisKey.EMOJI_KEY, emoji.id, payload.data.d.guild_id));
+                }
             }
         }
